+++ core/arch/arm/include/kernel/thread.h
@@ -218,13 +218,10 @@ typedef unsigned long (*thread_pm_handler_t)(unsigned long a0,
 					     unsigned long a1);
 struct thread_handlers {
 	/*
-	 * stdcall and fastcall are called as regular functions and
+	 * fastcall is called as a regular function and
 	 * normal ARM Calling Convention applies. Return values are passed
 	 * args->param{1-3} and forwarded into r0-r3 when returned to
 	 * non-secure world.
-	 *
-	 * stdcall handles calls which can be preemted from non-secure
-	 * world. This handler is executed with a large stack.
 	 *
 	 * fastcall handles fast calls which can't be preemted. This
 	 * handler is executed with a limited stack. This handler must not
@@ -235,7 +232,6 @@ struct thread_handlers {
 	 * on different stacks allowing native interrupts to be enabled during
 	 * a fastcall.
 	 */
-	thread_smc_handler_t std_smc;
 	thread_smc_handler_t fast_smc;
 
 	/*
