+++ core/arch/arm/kernel/thread_optee_smc.c
@@ -5,6 +5,7 @@
 
 #include <assert.h>
 #include <compiler.h>
+#include <io.h>
 #include <kernel/misc.h>
 #include <kernel/msg_param.h>
 #include <kernel/thread.h>
@@ -15,6 +16,7 @@
 #include <sm/optee_smc.h>
 #include <sm/sm.h>
 #include <string.h>
+#include <tee/entry_std.h>
 #include <tee/tee_cryp_utl.h>
 #include <tee/tee_fs_rpc.h>
 
@@ -92,6 +94,104 @@ static void thread_rpc_free_arg(uint64_t cookie)
 	}
 }
 
+static struct mobj *get_cmd_buffer(paddr_t parg, uint32_t *num_params)
+{
+	struct optee_msg_arg *arg;
+	size_t args_size;
+
+	arg = phys_to_virt(parg, MEM_AREA_NSEC_SHM);
+	if (!arg)
+		return NULL;
+
+	*num_params = READ_ONCE(arg->num_params);
+	if (*num_params > OPTEE_MSG_MAX_NUM_PARAMS)
+		return NULL;
+
+	args_size = OPTEE_MSG_GET_ARG_SIZE(*num_params);
+
+	return mobj_shm_alloc(parg, args_size, 0);
+}
+
+#ifdef CFG_CORE_DYN_SHM
+static struct mobj *map_cmd_buffer(paddr_t parg, uint32_t *num_params)
+{
+	struct mobj *mobj;
+	struct optee_msg_arg *arg;
+	size_t args_size;
+
+	assert(!(parg & SMALL_PAGE_MASK));
+	/* mobj_mapped_shm_alloc checks if parg resides in nonsec ddr */
+	mobj = mobj_mapped_shm_alloc(&parg, 1, 0, 0);
+	if (!mobj)
+		return NULL;
+
+	arg = mobj_get_va(mobj, 0);
+	if (!arg)
+		goto err;
+
+	*num_params = READ_ONCE(arg->num_params);
+	if (*num_params > OPTEE_MSG_MAX_NUM_PARAMS)
+		goto err;
+
+	args_size = OPTEE_MSG_GET_ARG_SIZE(*num_params);
+	if (args_size > SMALL_PAGE_SIZE) {
+		EMSG("Command buffer spans across page boundary");
+		goto err;
+	}
+
+	return mobj;
+err:
+	mobj_free(mobj);
+	return NULL;
+}
+#else
+static struct mobj *map_cmd_buffer(paddr_t pargi __unused,
+				   uint32_t *num_params __unused)
+{
+	return NULL;
+}
+#endif /*CFG_CORE_DYN_SHM*/
+
+static void std_smc_entry(struct thread_smc_args *smc_args)
+{
+	paddr_t parg = 0;
+	struct optee_msg_arg *arg = NULL;
+	uint32_t num_params = 0;
+	struct mobj *mobj = NULL;
+
+	if (smc_args->a0 != OPTEE_SMC_CALL_WITH_ARG) {
+		EMSG("Unknown SMC 0x%" PRIx64, (uint64_t)smc_args->a0);
+		DMSG("Expected 0x%x", OPTEE_SMC_CALL_WITH_ARG);
+		smc_args->a0 = OPTEE_SMC_RETURN_EBADCMD;
+		return;
+	}
+	parg = (uint64_t)smc_args->a1 << 32 | smc_args->a2;
+
+	/* Check if this region is in static shared space */
+	if (core_pbuf_is(CORE_MEM_NSEC_SHM, parg,
+			 sizeof(struct optee_msg_arg))) {
+		mobj = get_cmd_buffer(parg, &num_params);
+	} else {
+		if (parg & SMALL_PAGE_MASK) {
+			smc_args->a0 = OPTEE_SMC_RETURN_EBADADDR;
+			return;
+		}
+		mobj = map_cmd_buffer(parg, &num_params);
+	}
+
+	if (!mobj || !ALIGNMENT_IS_OK(parg, struct optee_msg_arg)) {
+		EMSG("Bad arg address 0x%" PRIxPA, parg);
+		smc_args->a0 = OPTEE_SMC_RETURN_EBADADDR;
+		mobj_free(mobj);
+		return;
+	}
+
+	arg = mobj_get_va(mobj, 0);
+	assert(arg && mobj_is_nonsec(mobj));
+	smc_args->a0 = tee_entry_std(arg, num_params);
+	mobj_free(mobj);
+}
+
 /*
  * Helper routine for the assembly function thread_std_smc_entry()
  *
@@ -103,7 +203,7 @@ void __weak __thread_std_smc_entry(struct thread_smc_args *args)
 #ifdef CFG_VIRTUALIZATION
 	virt_on_stdcall();
 #endif
-	thread_std_smc_handler_ptr(args);
+	std_smc_entry(args);
 
 	if (args->a0 == OPTEE_SMC_RETURN_OK) {
 		struct thread_ctx *thr = threads + thread_get_id();
