+++ core/arch/arm/kernel/thread_private.h
@@ -121,7 +121,6 @@ struct thread_ctx {
 extern const void *stack_tmp_export;
 extern const uint32_t stack_tmp_stride;
 extern struct thread_ctx threads[];
-extern thread_smc_handler_t thread_std_smc_handler_ptr;
 extern thread_smc_handler_t thread_fast_smc_handler_ptr;
 extern thread_nintr_handler_t thread_nintr_handler_ptr;
 extern thread_pm_handler_t thread_cpu_on_handler_ptr;
