+++ core/arch/arm/plat-bcm/main.c
@@ -19,7 +19,6 @@
 static void secure_intr_handler(void);
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = secure_intr_handler,
 	.cpu_on = cpu_on_handler,
