+++ core/arch/arm/plat-imx/main.c
@@ -51,7 +51,6 @@ static void main_fiq(void);
 static struct gic_data gic_data;
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 #if defined(CFG_WITH_ARM_TRUSTED_FW)
