+++ core/arch/arm/plat-k3/main.c
@@ -53,8 +53,6 @@ static void main_fiq(void)
 }
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
-	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = cpu_on_handler,
 	.cpu_off = pm_do_nothing,
