+++ core/arch/arm/plat-poplar/main.c
@@ -21,7 +21,6 @@
 static void main_fiq(void);
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = cpu_on_handler,
