+++ core/arch/arm/plat-rcar/main.c
@@ -54,7 +54,6 @@ register_dynamic_shm(NSEC_DDR_3_BASE, NSEC_DDR_3_SIZE);
 static void main_fiq(void);
 
 static const struct thread_handlers handlers __nex_data = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = cpu_on_handler,
