+++ core/arch/arm/plat-rockchip/main.c
@@ -51,7 +51,6 @@ static void main_fiq(void)
 }
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = pm_do_nothing,
