+++ core/arch/arm/plat-sam/main.c
@@ -53,7 +53,6 @@ static void main_fiq(void)
 }
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = pm_panic,
