+++ core/arch/arm/plat-sprd/main.c
@@ -51,7 +51,6 @@ register_phys_mem_pgdir(MEM_AREA_IO_SEC,
 static void main_fiq(void);
 
 static const struct thread_handlers handlers = {
-	.std_smc = tee_entry_std,
 	.fast_smc = tee_entry_fast,
 	.nintr = main_fiq,
 	.cpu_on = cpu_on_handler,
